move LabelerConfig to appropriate package level
